test: relocate confidence-band tests to test_config_inbox.py per project pattern
