+++ tests/moneybin/test_config.py
@@ -81,57 +81,3 @@ def test_ai_config_env_override(monkeypatch: pytest.MonkeyPatch) -> None:
     settings = MoneyBinSettings()
     assert settings.ai.default_backend == "anthropic"
     assert settings.ai.consent_policy == "strict"
-
-
-class TestConfidenceBands:
-    """Validate confidence band thresholds and constraints."""
-
-    def test_defaults_per_spec(self) -> None:
-        from moneybin.config import ConfidenceBands
-
-        b = ConfidenceBands()
-        assert b.t_high == 0.90
-        assert b.t_med == 0.70
-
-    def test_rejects_inverted(self) -> None:
-        from moneybin.config import ConfidenceBands
-
-        with pytest.raises(ValueError, match="t_high must be >= t_med"):
-            ConfidenceBands(t_high=0.5, t_med=0.7)
-
-    def test_rejects_out_of_range(self) -> None:
-        from moneybin.config import ConfidenceBands
-
-        with pytest.raises(ValueError):
-            ConfidenceBands(t_high=1.5, t_med=0.7)
-        with pytest.raises(ValueError):
-            ConfidenceBands(t_high=0.9, t_med=-0.1)
-
-    def test_accepts_equal_bands(self) -> None:
-        from moneybin.config import ConfidenceBands
-
-        b = ConfidenceBands(t_high=0.8, t_med=0.8)
-        assert b.t_high == b.t_med
-
-
-class TestImportSettingsConfidence:
-    """Validate ImportSettings carries confidence + self_accept_high."""
-
-    def test_default_self_accept_high_false(self) -> None:
-        from moneybin.config import ImportSettings
-
-        s = ImportSettings()
-        assert s.self_accept_high is False
-
-    def test_self_accept_high_can_enable(self) -> None:
-        from moneybin.config import ImportSettings
-
-        s = ImportSettings(self_accept_high=True)
-        assert s.self_accept_high is True
-
-    def test_confidence_bands_default(self) -> None:
-        from moneybin.config import ImportSettings
-
-        s = ImportSettings()
-        assert s.confidence.t_high == 0.90
-        assert s.confidence.t_med == 0.70
